remove mkTuple
