+++ core/amber/src/test/scala/edu/uci/ics/amber/engine/architecture/messaginglayer/NetworkInputGatewaySpec.scala
@@ -1,8 +1,9 @@
 package edu.uci.ics.amber.engine.architecture.messaginglayer
 
 import edu.uci.ics.amber.engine.common.ambermessage.{DataFrame, WorkflowFIFOMessage}
+import edu.uci.ics.amber.engine.common.tuple.amber.TupleLike
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, ChannelIdentity}
-import edu.uci.ics.amber.engine.utils.TupleFactory.mkTuple
+import edu.uci.ics.texera.workflow.common.tuple.schema.{AttributeType, Schema}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.flatspec.AnyFlatSpec
 
@@ -12,7 +13,14 @@ class NetworkInputGatewaySpec extends AnyFlatSpec with MockFactory {
   private val fakeSenderID = ActorVirtualIdentity("testSender")
   private val channelId = ChannelIdentity(fakeSenderID, fakeReceiverID, false)
   private val payloads = (0 until 4).map { i =>
-    DataFrame(Array(mkTuple(i)))
+    DataFrame(
+      Array(
+        TupleLike.enforceSchema(
+          TupleLike(i),
+          Schema.newBuilder().add("field1", AttributeType.INTEGER).build()
+        )
+      )
+    )
   }.toArray
   private val messages = (0 until 4).map { i =>
     WorkflowFIFOMessage(channelId, i, payloads(i))
