+++ core/amber/src/test/scala/edu/uci/ics/amber/engine/faulttolerance/LoggingSpec.scala
@@ -17,6 +17,7 @@ import edu.uci.ics.amber.engine.common.ambermessage.{
 }
 import edu.uci.ics.amber.engine.common.rpc.AsyncRPCClient.ControlInvocation
 import edu.uci.ics.amber.engine.common.storage.SequentialRecordStorage
+import edu.uci.ics.amber.engine.common.tuple.amber.TupleLike
 import edu.uci.ics.amber.engine.common.virtualidentity.{
   ActorVirtualIdentity,
   ChannelIdentity,
@@ -25,7 +26,7 @@ import edu.uci.ics.amber.engine.common.virtualidentity.{
 }
 import edu.uci.ics.amber.engine.common.virtualidentity.util.{CONTROLLER, SELF}
 import edu.uci.ics.amber.engine.common.workflow.{PhysicalLink, PortIdentity}
-import edu.uci.ics.amber.engine.utils.TupleFactory.mkTuple
+import edu.uci.ics.texera.workflow.common.tuple.schema.{AttributeType, Schema}
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.flatspec.AnyFlatSpecLike
 
@@ -49,7 +50,21 @@ class LoggingSpec
     ControlInvocation(0, AddPartitioning(mockLink, mockPolicy)),
     ControlInvocation(0, PauseWorker()),
     ControlInvocation(0, ResumeWorker()),
-    DataFrame((0 to 400).map(i => mkTuple(i, i.toString, i.toDouble)).toArray),
+    DataFrame(
+      (0 to 400)
+        .map(i =>
+          TupleLike.enforceSchema(
+            TupleLike(i, i.toString, i.toDouble),
+            Schema
+              .newBuilder()
+              .add("field1", AttributeType.INTEGER)
+              .add("field2", AttributeType.STRING)
+              .add("field3", AttributeType.DOUBLE)
+              .build()
+          )
+        )
+        .toArray
+    ),
     ControlInvocation(0, StartWorkflow()),
     ControlInvocation(0, WorkerExecutionCompleted())
   )
